class Chunk
Stringstream-like thing for formatted output to a buffer.

The similar but more interesting ResponseBuf got removed with this
commit.  It came from the aborted NNTP work way back, and kind of
combined Chunk, filter and backlog.

Not directly useful here though, since it assumed formatting a complete
response and then writing it makes sense. That is true for NNTP, but
not for HTTP (e.g. an ISO image is huge and doesn't need formatting).