thumbnailing: don't create empty files
cjpeg(1) tends to create its output file even if it fails,
so we have to delete it in that case -- or we'd start serving
empty files as thumbnail JPEGs.

This fix has the negative effect of defeating "negative caching":
N requests for a problem thumbnail will make us try to generate
it N times.