css for galleries
Galleries as a series of fixed-size <div>s. Seems ok, except I
haven't managed to center the thumbnails in the div yet, and I
haven't solved having text alongside the thumbnail.

And in a way it would be nicer to follow the timeline down the
columns instead of through the rows like text.