Request: fix case sensitivity
Header names and "HTTP" are case-insensitive; "GET" is not.

At lookup, try a case-sensitive match against the "canonical"
name first, and only do a case-insensitive if that fails.
That gives an attacker a way to make her requests more expensive,
but not by much.  Typical clients will not hit the second code
path.